PartyRecomender/recommender/requirements.txt
fastapi==0.68.0
tekore==3.7.1
uvicorn==0.15.0
pandas==1.3.2
flask==2.0.1
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
lightfm==1.16
python-dotenv==0.19.0